Rename MDAPoint reducer and actions page name
